完成单个area光源的实现

+++ Model.h
@@ -19,7 +19,7 @@ typedef struct Mesh
 	std::vector<Vertex> vertices;
 	std::vector<Index> indices;
 	XMFLOAT3 center;
-	double surfaceArea;
+	float surfaceArea;
 };
 
 typedef struct Model
